Support for switching between chips. The new file overlock_chip_support contains the code which runs at boot. However it is called from set-kano-hdmi to avoid an additional reboot.
When the other chip is detected, the existing clock values are saved to
one of two backup config files /boot/config_pi{1,2}_backup.txt

If a backup file exists for teh new chip type and contains valid clock values, these
are restored. Otherwise default lock values are selected.

overclock.py contains some new support functions, for saving overlock values
to backup config files.

Note that it is not smart about checking compatibility with the chip, so
if we tweak the values at all in future versions, people will get reverted to the default.

Also it is notable that logs get lost when the reboot happens, but I have not fixed this
since it will require updating kano-toolset, so should be done when we are going to rev that anyway.