Parental: Blacklist Google webcache and translate
KanoComputing/peldins#1692
Cached versions of blocked sites can be seen either by viewing Google's
webcache of the site or by translating the URL so block this in the
`/etc/hosts` blacklist mode of parental control. Do this by creating a
Kano list for blacklisting to supplement that supplied by
www.hostsfile.org/hosts.html. Updates the instructions in the  LICENSE
readme to reflect this extra step.