i18n: Add functionality (and tests) to set locale
To enable use of locales, add the functions which provide the ability to
set them using the settings toolset. This adds a functions to manage the
locales (installation, removal, setting the locale, checking that the
locale is valid, etc) and also provides unit tests to check each of
these functions.