Parental: Add function to retrieve parental level
KanoComputing/peldins#1919
Although the parental level can be retrieved through `config_file`'s
`get_setting` function, this is intended for use solely within Kano
Settings (for the UI). Also, to get the correct setting, a combination
of `get_setting` and checking if the parental controls are enabled at
all is required to get the correct setting so add a function to allow
the setting to be retrieved by third parties.