Parental: Prevent hanging when starting sentry server
KanoComputing/peldins#2175
When attempting to connect to WiFi, with ultimate parental controls
enabled and the sentry server not running, the GUI can hang despite
having fully established a connection. This is because launching the
sentry server was done in a way that it was waiting on stdin. Resolve
this by explicitly declaring the stdin source.