Boot Config: Fix tests failing due to equivalent changes
When parsing the configuration file and dumping the data again, all
empty comment lines will be reduced to an empty line for simplicity. The
test case had instances of such a comment which meant that the test
would fail when comparing the two equivalent lines. Prevent it from
failing because of this.