Fix for unresponsive flip display option in Display
Using an RPi2 with a Screen Kit, it was possible somehow for something to get
confused about the display and write the option at the end. What that means is
that kano-settings will change the display_rotate in one of the Screen Kit
filters which gets ignored by the lowest setting that was incorrectly writen
at the end of the file outside the filters.