Fix Audio screen to show available features and system status
The change came from having to disable jack audio because of PiHat.
The radio buttons get disabled if audio through HDMI or JACK are
not supported. Also, now they show what is currently set on the
system. One risk of disabling the HDMI audio button is that EDID
info may be full of lies, in which case users might need to force
the setting.