Boot Config: Default "hdmi_drive" to 2
The `hdmi_drive` setting specifies whether the HDMI port should drive
the display with HDMI or DVI signalling for DMT modes. If set to DVI
signalling, then no sound can be transmitted over the wire and hence TVs
set to DMT modes will not have any sound in this case. When set to HDMI
signalling, if sound is supported then it will be sent with the video.

There is potential for problems if the signalling is set to HDMI on a
monitor which is set to a DMT resolution, supports audio and doesn't
support HDMI signalling; in this case there will be no sound and
picture, however these monitors seem very rare.

More information can be found in the docs:
    http://elinux.org/RPiconfig
and a very interesting forum thread containing more detailed information
about the subject:
    https://www.raspberrypi.org/forums/viewtopic.php?f=66&t=169879