Moved old tests to pytest with Make and added tests for Display
This commit is two fold. Firstly, all the old tests were moved
under pytest with a standard Makefile to run them. Here, the
TestConfig class was disabled because it required sudo permissions.
Secondly, tests for the automatic resolution policy were added.
These use fixtures to load a dataset of screen dumps and run check
that the display functions return & apply the correct settings.