Added display configurations for all SKs to avoid reboots
In order to avoid having to detect the optimal resolution for all
our SKs and then reboot, we explicitly set the modes in the config
file. The default hdmi_group/mode have also been set for an uSK.
Lastly, the hdmi_drive=2 option was removed in order to avoid
forcing some monitors to output an unsupported signal.