Added EDID tools for dumping and flashing screen data
Two tools have been added here: rpi-dump-edid is a dev & user facing
script which dumps the screen EDID to the /boot/ partition for testing
data and debugging. generate-rpi-edid-dumps is a tool which uses raw
EDIDs to generate test data for kano-settings display functions.
These will be the tools to build on the res/edid dataset.