Fixed resetting Audio > Boost Volume setting on CKCv1
The issue here was that on CKCv1, kano-settings-onboot script set
the volume to low every time. The user is able to raise the volume
through Boost Volume option in the Audio menu which the bootup
script was not aware of, resetting the config. Created a system
file flag to note that the volume was lowered only once and the
user is responsible for the volume level thereafter.