Backported v3.13.0 /boot/config.txt to the v3.9.2 es_AR image
The reson for this being that the automatic resolution configuration
policy set a narrower resolution on one of our SKs. This caused the
Argentinian wallpaper banner to be off the screen and not visible
from Dashboard. Together with this, the new config file also contains
a disabled RPi warnings option, which will remove the underpowered
and thermal warnings (kit will have a battery).