Parental: Direct the whitelist to the master branch
The ultimate parental settings refer to the online URL of the whitelist
file, however this reference is from a time when the `jessie` branch was
the primary branch. This means that all changes to this file in `master`
will not propagate to these users. Migrate this over to the `master`
branch so at least all up-to-date customers will receive the correct
update to their settings.