Tests: Fix using open() in fixtures with fake fs
When integrating fixtures which read data from the file system with the
fake file system fixture, often the case can occur where the fixture
which attempts to read file data attempts to do it from the (empty) fake
file system rather than the real one where the file is actually located.
Mitigate this problem for fixtures which suffer from this problem.